• 0.3.3:     - Fix bug in info.py importation     - Major restructuring of common.py
