+++ src/modules/dependencies.py
@@ -1,0 +1,22 @@
+# OS modules
+import os
+import sys
+import importlib
+import subprocess
+
+# Time and date
+import time
+import datetime
+
+
+# Data manipulation
+import ltspice
+import h5py
+import numpy as np
+import pandas as pd
+from scipy.interpolate import interp1d
+
+
+# Data visualization
+import matplotlib.pyplot as plt
+from matplotlib.animation import FuncAnimation
