--- test.py
@@ -1,1 +1,0 @@
-for i in range(10 - 1, -1, -1): print(i)
